move batch_dedup, reactor, file_descriptor
